addMetrics() annotates a span with URLSessionTaskMetrics
This was already provided by urlSession(_:task:didFinishCollecting:),
but there are cases where it's more convenient to annotate the span
directly without providing the 'session' and 'task' objects (neither of
which is needed to set a span's metrics attributes.

addMetrics() is called from urlSession(_:task:didFinishCollecting:) so
backwards compatibility is preserved.