startSpan(for: URLRequest) sets all request attributes
Set the full set of URLRequest-derived span attributes. These were
previously only set by urlSession(_:didCreateTask:), but we have all
of the context we need to set them here.

This makes urlSession(_:didCreateTask:) entirely optional to call when a
span is constructed from a URLRequest. Doing so will result in the
delegate call replacing/updating the values added by startSpan(), which
might be useful if some URLRequest properties (such as headers) were set
in the interim.