Improve startSpan argument names
Remove the 'for' label name in favor of 'request'. This feels more
natural and obvious given that its an `inout` parameter.

Also, prefer 'captureHeaders' to 'headersToCapture' to communicate this
is an action that this method will perform; that set isn't retained and
won't be applied to any future headers that are added to the request.

Lastly, we had some redundant tests for this function, so consolidate
them.