Introduce URLTemplateMatcher
URLTemplateMatcher matches URLs against a collection of URL template
patterns. Template patterns describe a URL path and query string. They
can use templated parameters like {} or {id} to match dynamic portions
of the path and query parameter values ("/users/{id}?status={status}").

This is useful for creating an explicit list of possible URL template
values and then selecting the best match for a given URL. This template
value can then be used to to construct the Span; the template will be
used to form the Span's name and added as the "url.template" attribute.